MoveForm/requirements.txt
django
Django
django_bootstrap3
django_bootstrap4
django_crispy_forms
pip
pytz

setuptools

coverage
twine
wheel
pytest
pytest-cov
pytest-django